requirements.txt
diff-match-patch==20200713
dj-database-url==0.4.2
Django==1.11.17
django-crispy-forms==1.10.0
gunicorn==19.7.1
Pillow==8.0.1
psycopg2==2.7.1
psycopg2-binary==2.8.6
pytz==2020.4
whitenoise==3.3.0
